+++ SpringProjects/FullCrud/src/main/webapp/statments.jsp
@@ -12,22 +12,15 @@
 <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 </head>
 <body>
-	<div class="container">
-	
-		<div class="row">
-			<div class="col">		
-				<h1><c:out value="${book.title}"/></h1>
-			</div>
-			<div class="col">
-				<a href="/books"> back to the shelves</a>
-			</div>	
-		</div>
-	</div>
 
-	<div class="container">	
+
+C:OUT
+			<c:out value="${book.description}"/>
+
+
+CHOOSE STATEMENT
 		<div>
 			<p>
-		<!-- 	// THIS CODE ALLOW YOU TO SHOW A USER WHO IS LOGGED IN ONE THING COMPARED TO IF THEY'RE NOT LOGGED IN -->
 			<c:choose>
 				<c:when test = "${user_id == book.creator.id }">
 					You read <c:out value="${book.title}"/>
@@ -45,22 +38,20 @@
 						
 			</p>
 		</div>
-<%-- 		<div>
-			<p>
-				Here are <c:out value="${book.creator.firstName}"/>'s Thoughts
-			</p>
-		</div> --%>
-		<div>
-			<p>
-				<c:out value="${book.description}"/>
-			</p>
-		</div>
+		
+IF STATEMENT
+
 		<c:if test="${user_id == book.creator.id }">
 			<a class="btn btn-dark" href="/books/${book.id}/edit">Edit</a>
 			<a class="btn btn-dark" href="/book/${book.id}/delete ">Delete</a>
 		</c:if>
 		
-	</div>
+		
+DELETE
+		
+		
+	<a class="btn btn-dark" href="/book/${book.id}/delete ">Delete</a>	
+		
 
 
 <!-- JavaScript Bundle with Popper -->
